Introduced error module with hasError + moved existing error functions
